cli: split sparse sub commands into modules
